refactor(security): rename LoginUser to AuthUserDetails and update related components
- Renamed LoginUser class to AuthUserDetails
- Updated all references from LoginUser to AuthUserDetails
- Renamed CurrentUser annotation to AuthUser
- Renamed CurrentUserMethodArgumentResolver to AuthUserMethodArgumentResolver
- Updated method signatures and variable names accordingly
- Adjusted type parameters and casting in security-related services
- Modified Redis key definitions to use AuthUserDetails
- Updated controller method parameters to use AuthUser annotation
- Changed return types in security service interfaces
- Updated implementation classes to reflect new naming conventions

+++ migoo-webserver-parent/migoo-webserver-infra-station-letter/src/main/java/xyz/migoo/framework/infra/service/StationLetterServiceImpl.java
@@ -32,7 +32,7 @@ public class StationLetterServiceImpl implements StationLetterService {
     public Long add(String code, String title, String content, String fromUserType, Long toUserId, String toUserType) {
         StationLetterDO bean = new StationLetterDO().setCode(code).setTitle(title).setContent(content)
                 .setToUserId(toUserId).setToUserType(toUserType)
-                .setFromUserId(SecurityFrameworkUtils.getLoginUserId()).setFromUserType(fromUserType);
+                .setFromUserId((Long) SecurityFrameworkUtils.getLoginUserId()).setFromUserType(fromUserType);
         mapper.insert(bean);
         return bean.getId();
     }
